# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HeartSpace Counselling Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type LoginBodyRole = (typeof LoginBodyRole)[keyof typeof LoginBodyRole];

export const LoginBodyRole = {
  student: "student",
  counsellor: "counsellor",
} as const;

export interface LoginBody {
  email: string;
  password: string;
  role: LoginBodyRole;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  student: "student",
  counsellor: "counsellor",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type SessionStatus = (typeof SessionStatus)[keyof typeof SessionStatus];

export const SessionStatus = {
  scheduled: "scheduled",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Session {
  id: number;
  studentId: number;
  counsellorId: number;
  student?: User;
  counsellor?: User;
  scheduledAt: string;
  durationMinutes: number;
  status: SessionStatus;
  notes?: string | null;
  topic?: string | null;
  createdAt: string;
}

export interface CreateSessionBody {
  studentId: number;
  counsellorId: number;
  scheduledAt: string;
  durationMinutes: number;
  topic?: string;
}

export type UpdateSessionBodyStatus =
  (typeof UpdateSessionBodyStatus)[keyof typeof UpdateSessionBodyStatus];

export const UpdateSessionBodyStatus = {
  scheduled: "scheduled",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateSessionBody {
  status?: UpdateSessionBodyStatus;
  notes?: string;
  scheduledAt?: string;
}

export interface MoodEntry {
  id: number;
  studentId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  mood: number;
  note?: string | null;
  createdAt: string;
}

export interface CreateMoodBody {
  studentId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  mood: number;
  note?: string;
}

export interface DashboardSummary {
  totalSessions: number;
  upcomingSessions: number;
  completedSessions: number;
  averageMood?: number | null;
  recentSessions: Session[];
  totalStudents: number;
}

export interface StudentOverview {
  student: User;
  totalSessions: number;
  lastSession?: string | null;
  latestMood?: number | null;
  upcomingSession?: Session | null;
}

export type ListUsersParams = {
  role?: ListUsersRole;
  search?: string;
};

export type ListUsersRole = (typeof ListUsersRole)[keyof typeof ListUsersRole];

export const ListUsersRole = {
  student: "student",
  counsellor: "counsellor",
} as const;

export type ListSessionsParams = {
  studentId?: number;
  status?: ListSessionsStatus;
};

export type ListSessionsStatus =
  (typeof ListSessionsStatus)[keyof typeof ListSessionsStatus];

export const ListSessionsStatus = {
  scheduled: "scheduled",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type ListMoodsParams = {
  studentId?: number;
  limit?: number;
};

export type GetDashboardSummaryParams = {
  studentId?: number;
};