# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HeartSpace Counselling Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Session } from "./session";

export interface DashboardSummary {
  totalSessions: number;
  upcomingSessions: number;
  completedSessions: number;
  averageMood?: number | null;
  recentSessions: Session[];
  totalStudents: number;
}