# lib/api-zod/src/generated/types/moodEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HeartSpace Counselling Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface MoodEntry {
  id: number;
  studentId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  mood: number;
  note?: string | null;
  createdAt: Date;
}