# lib/api-zod/src/generated/types/studentOverview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HeartSpace Counselling Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Session } from "./session";
import type { User } from "./user";

export interface StudentOverview {
  student: User;
  totalSessions: number;
  lastSession?: Date | null;
  latestMood?: number | null;
  upcomingSession?: Session | null;
}