# lib/api-zod/src/generated/types/updateSessionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HeartSpace Counselling Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateSessionBodyStatus } from "./updateSessionBodyStatus";

export interface UpdateSessionBody {
  status?: UpdateSessionBodyStatus;
  notes?: string;
  scheduledAt?: Date;
}